requirements.txt
discord.py>=2.3.2
aiohttp>=3.9.0
python-dotenv>=1.0.0
asyncio>=3.4.3
Pillow>=10.0.0
